farmer/재배방법/pyproject.toml
[project]
name = "finalproject"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12.4,<4.0"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "langchain-openai (>=0.3.21,<0.4.0)",
    "gradio (>=5.33.1,<6.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "langchain-google-genai (>=2.1.5,<3.0.0)",
    "langchain-community (>=0.3.25,<0.4.0)",
    "faiss-cpu (>=1.11.0,<2.0.0)",
    "pypdf (>=5.6.0,<6.0.0)",
    "langchain-huggingface (>=0.3.0,<0.4.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "transformers (>=4.52.4,<5.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "docx2txt (>=0.9,<0.10)",
    "langgraph (>=0.4.8,<0.5.0)",
    "pdfplumber (>=0.11.7,<0.12.0)",
    "streamlit (>=1.45.1,<2.0.0)",
    "langchain-experimental (>=0.3.4,<0.4.0)",
    "langchain-ollama (>=0.3.3,<0.4.0)",
    "langchain-tavily (>=0.2.2,<0.3.0)",
    "wikipedia (>=1.4.0,<2.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "tabulate (>=0.9.0,<0.10.0)",
    "langserve (>=0.3.1,<0.4.0)",
    "fastapi (>=0.115.14,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "sse-starlette (>=2.3.6,<3.0.0)",
    "langsmith (>=0.4.3,<0.5.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

